backend/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
qdrant-client==1.7.0
openai==0.27.10
python-dotenv==1.0.1
httpx==0.26.0
better-auth==0.0.1b11